Move packages inside src
