--- packages/excalidraw/index.tsx
@@ -1,3 +1,0 @@
-import ExcalidrawApp from "../../src/excalidraw-app/app";
-
-export default ExcalidrawApp;
