+++ src/packages/excalidraw/webpack.prod.config.js
